requirements.txt
selenium==4.15.0
pytest==7.4.3
requests==2.31.0
allure-pytest==2.13.2
pytest-html==4.1.1
webdriver-manager==4.0.1
python-dotenv==1.0.0